tarea1 parte2

+++ kernel/proc.h
@@ -107,4 +107,4 @@ struct proc {
 };
  
 uint getppid(void);
-int getancestor(int);
+int getancestor(uint64);
